services/server/requirements.txt
Flask
Flask-SQLAlchemy
PyJWT==1.7.1
Werkzeug
python-whois
waitress
psycopg2-binary==2.8.6
